Move Akka-independent execution to own module
