+++ akka/src/main/scala/io/kagera/akka/query/PetriNetQuery.scala
@@ -7,9 +7,8 @@ import akka.stream.scaladsl._
 import io.kagera.akka.actor.{ AkkaObjectSerializer, PetriNetInstance }
 import io.kagera.api.colored.ExecutablePetriNet
 import io.kagera.execution.EventSourcing._
-import io.kagera.execution._
+import io.kagera.execution.Instance
 import io.kagera.persistence.Serialization
-import io.kagera.persistence.Serialization._
 
 trait PetriNetQuery[S] {
